requirements.txt
asgiref==3.6.0
Authlib==1.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
phonenumbers==8.13.7
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
